Fix cleaner to delete table only if it exists

+++ app/dynamo_db/cleaner.rb
@@ -4,9 +4,12 @@ module DynamoDatabase
     include Connection
 
     def clean
+      table_names = client.list_tables.table_names
       Initializer.load_table_classes.each do |clazz|
         begin
-          client.delete_table(table_name: clazz.get_table_name)
+          if table_names.include? clazz.get_table_name
+            client.delete_table(table_name: clazz.get_table_name)
+          end
         rescue Aws::DynamoDB::Errors::LimitExceededException
           sleep 1
           retry
